+++ oneview_redfish_toolkit/api/computer_system_collection.py
@@ -23,28 +23,30 @@ class ComputerSystemCollection(RedfishJsonValidator):
     """Creates a Computer System Collection Redfish dict
 
         Populates self.redfish with some hardcoded ComputerSystemCollection
-        values and with the response of Oneview with all server hardware.
+        values and with the response of Oneview with all servers profiled
+        applied.
     """
 
     SCHEMA_NAME = 'ComputerSystemCollection'
 
-    def __init__(self, server_hardware):
+    def __init__(self, server_profiles):
         """ComputerSystemCollection constructor
 
             Populates self.redfish with a hardcoded ComputerSystemCollection
-            values and with the response of Oneview with all server hardware.
+            values and with the response of Oneview with all servers profiled
+            applied.
 
             Args:
-                server_hardware: A list of dicts of server hardware.
+                server_profiles: A list of dicts of servers profiled applied.
         """
         super().__init__(self.SCHEMA_NAME)
 
-        self.server_hardware = server_hardware
+        self.server_profiles = server_profiles
 
         self.redfish["@odata.type"] = \
             "#ComputerSystemCollection.ComputerSystemCollection"
         self.redfish["Name"] = "Computer System Collection"
-        self.redfish["Members@odata.count"] = len(server_hardware)
+        self.redfish["Members@odata.count"] = len(server_profiles)
         self.redfish["Members"] = list()
         self._set_redfish_members()
         self.redfish["@odata.context"] = \
@@ -59,10 +61,10 @@ class ComputerSystemCollection(RedfishJsonValidator):
             Populates self.redfish["Members"] with the links to Redfish
             ComputerSystems.
         """
-        for server_hardware, index in \
-                zip(self.server_hardware,
-                    range(len(self.server_hardware))):
-
+        for server_profile, index in \
+                zip(self.server_profiles, range(len(self.server_profiles))):
             self.redfish["Members"].append(collections.OrderedDict())
+            server_profile_uuid = \
+                server_profile["serverProfileUri"].split("/")[-1]
             self.redfish["Members"][index]["@odata.id"] = \
-                "/redfish/v1/Systems/" + server_hardware["uuid"]
+                "/redfish/v1/Systems/" + server_profile_uuid